* moved prepare_randomized_search to it's own file
